Crate oak_functions

SDK functionality that provides idiomatic Rust wrappers around the underlying Oak Functions platform functionality.

Functions

alloc

read_request
Reads and returns the user request.

result_from_status
Convert a status returned from a host function call to a Result.

storage_get_item
Looks up an item from the in-memory lookup store.

write_response
Write the response.