Function oak_functions::report_metric

pub fn report_metric<T: AsRef<str>>(
    label: T, 
    value: i64
) -> Result<(), OakStatus>

Reports a metric value for a sum-based metrics bucket.

If differentially-private metrics are enabled in the configuration the metrics bucket totals will be logged in batches after sufficient noise has been added. If multiple values are reported for the same bucket during a single request only the last value will be used. If the reported value falls outside of the configured range for the bucket, it will be clamped to the minimum or maximum value.

If no values are reported for a configured bucket during a request, it will be treated as if 0 was reported. If the minimum value of the bucket is larger than 0 it would then be clamped to the configured minimum. This could lead to unexpected bias in the results, so minimum values above 0 should be used with care.

See report_metric.